Extract signing_time_is_valid method to utils.
